add a single constant file for frontend
