+++ Frontend/src/components/NotesCard.jsx
@@ -1,4 +1,4 @@
-import { NOTES_BUTTON } from "@/constants/notes";
+import { NOTES_BUTTON } from "@/constants/constants";
 import React from "react";
 import { Link } from "react-router-dom";
 import { Button } from "./ui/button";
