+++ Frontend/src/components/PageLayout.jsx
@@ -1,12 +1,15 @@
-import { Outlet } from "react-router-dom";
+import { Outlet, useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { useCallback, useEffect } from "react";
 import AuthService from "@/services/AuthService";
 import ApiError from "@/services/ApiError";
 import { login, logout, updateLoginCheckDone } from "@/store/AuthSlice";
+import { ROUTES } from "@/constants/route";
+import { toast } from "react-toastify";
 
 function PageLayout() {
   const dispatch = useDispatch();
+  const navigate = useNavigate();
 
   const fetchUser = useCallback(async () => {
     const response = await AuthService.getCurrentUser(
@@ -15,9 +18,16 @@ function PageLayout() {
         "Cache-Control": "no-cache, no-store, must-revalidate",
       }
     );
-
+    console.log(response);
     if (!(response instanceof ApiError)) {
       dispatch(login(response?.data));
+    } else if (
+      response instanceof ApiError &&
+      response?.errorResponse?.message === "Please verify your email"
+    ) {
+      dispatch(logout());
+      navigate(`${ROUTES.VERIFYCODE}`);
+      toast.error(response?.errorResponse?.message);
     } else {
       dispatch(logout());
     }
