+++ Frontend/src/components/Sidebar/Sidebar.jsx
@@ -3,7 +3,7 @@ import { useMediaQuery } from "usehooks-ts";
 import DesktopSidebar from "./DesktopSidebar";
 import MobileSidebar from "./MobileSidebar";
 import { ROUTES } from "../../constants/route";
-import { SIDEBAR_TABS } from "../../constants/sidebar";
+import { SIDEBAR_TABS } from "../../constants/constants";
 
 const sideBarItems = [
   { label: `${SIDEBAR_TABS.HOME}`, href: `${ROUTES.HOME}`, icon: Home },
