--- Frontend/src/constants/account.js
@@ -1,32 +1,0 @@
-// Account constants
-
-export const TITLE = "Account Page";
-
-export const TABS = {
-  DEFAULT_VALUE: "account",
-  ACCOUNT: "account",
-  PASSWORD: "password",
-};
-
-export const TEXT_CONTENT = {
-  ACCOUNT_TAB: "Account",
-  PASSWORD_TAB: "Password",
-  ACCOUNT_TITLE: "Account",
-  ACCOUNT_DESCRIPTION:
-    "Make changes to your account here. Click save when you're done.",
-  PASSWORD_TITLE: "Password",
-  PASSWORD_DESCRIPTION:
-    "Change your password here. After saving, you'll be logged out.",
-};
-
-export const UPDATE_SEMESTER_TEXT_CONTENT = {
-  TITLE: "Update Semester",
-  DESCRIPTION:
-    "Update your semester here. Click udpate button when you're done.",
-};
-
-export const BUTTONS = {
-  SAVE_CHANGES: "Save changes",
-  SAVE_PASSWORD: "Save password",
-  UPDATE_SEMESTER: "Update",
-};
