--- Frontend/src/constants/auth.js
@@ -1,35 +1,0 @@
-// Auth constant
-
-// Dialog and Form Titles
-export const TITLE = "Join Notes Library";
-export const INPUT_OTP_FORM_TITLE = "One-Time Password";
-export const SIGNUP_DESCRIPTION = "Sign up to deep dive into the notes world";
-export const SIGNIN_DESCRIPTION = "Sign in to deep dive into the notes world";
-export const INPUT_OTP_FORM_DESCRITION =
-  "Please enter the one-time password sent to your email.";
-export const RESEND_EMAIL_DESCRIPTION =
-  "Didn’t receive the verification email?";
-
-export const SEMESTER_OPTIONS = [
-  { id: 1, name: "Fifth" },
-  { id: 2, name: "Sixth" },
-];
-
-export const COURSE_OPTIONS = [
-  {
-    id: 1,
-    name: "B.Tech-CS",
-  },
-];
-
-// Button Text
-export const SIGNUP_BUTTON_TEXT = "Sign up";
-export const SIGNIN_BUTTON_TEXT = "Sign in";
-export const SIGNIN_PROMPT_TEXT = "Already a member?";
-export const SIGNUP_PROMPT_TEXT = "Don't have an account";
-export const SUBMIT_BUTTON = "Submit";
-export const RESEND_EMAIL_BUTTON_TEXT = "Resend Verification Email";
-export const USER_ROLE = {
-  ADMIN: "ADMIN",
-  USER: "USER",
-};
