+++ Frontend/src/constants/constants.js
@@ -1,0 +1,112 @@
+// Auth constant
+
+// Dialog and Form Titles
+export const TITLE = "Join Notes Library";
+export const INPUT_OTP_FORM_TITLE = "One-Time Password";
+export const SIGNUP_DESCRIPTION = "Sign up to deep dive into the notes world";
+export const SIGNIN_DESCRIPTION = "Sign in to deep dive into the notes world";
+export const INPUT_OTP_FORM_DESCRITION =
+  "Please enter the one-time password sent to your email.";
+export const RESEND_EMAIL_DESCRIPTION =
+  "Didn’t receive the verification email?";
+
+export const SEMESTER_OPTIONS = [
+  { id: 1, name: "Fifth" },
+  { id: 2, name: "Sixth" },
+];
+
+export const COURSE_OPTIONS = [
+  {
+    id: 1,
+    name: "B.Tech-CS",
+  },
+];
+
+// Button Text
+export const SIGNUP_BUTTON_TEXT = "Sign up";
+export const SIGNIN_BUTTON_TEXT = "Sign in";
+export const SIGNIN_PROMPT_TEXT = "Already a member?";
+export const SIGNUP_PROMPT_TEXT = "Don't have an account";
+export const SUBMIT_BUTTON = "Submit";
+export const RESEND_EMAIL_BUTTON_TEXT = "Resend Verification Email";
+export const USER_ROLE = {
+  ADMIN: "ADMIN",
+  USER: "USER",
+};
+
+// Account constants
+
+export const ACCOUNT_TITLE = "Account Page";
+
+export const TABS = {
+  DEFAULT_VALUE: "account",
+  ACCOUNT: "account",
+  PASSWORD: "password",
+};
+
+export const ACCOUNT_TEXT_CONTENT = {
+  ACCOUNT_TAB: "Account",
+  PASSWORD_TAB: "Password",
+  ACCOUNT_TITLE: "Account",
+  ACCOUNT_DESCRIPTION:
+    "Make changes to your account here. Click save when you're done.",
+  PASSWORD_TITLE: "Password",
+  PASSWORD_DESCRIPTION:
+    "Change your password here. After saving, you'll be logged out.",
+};
+
+export const UPDATE_SEMESTER_TEXT_CONTENT = {
+  TITLE: "Update Semester",
+  DESCRIPTION:
+    "Update your semester here. Click udpate button when you're done.",
+};
+
+export const BUTTONS = {
+  SAVE_CHANGES: "Save changes",
+  SAVE_PASSWORD: "Save password",
+  UPDATE_SEMESTER: "Update",
+};
+
+// Home constants
+
+export const TYPED_STRINGS = ["Efficient", "Easy", "and", "Affordable"];
+
+export const HOME_TEXT_CONTENT = {
+  MAIN_TITLE: "Stay Ahead Of The Curve",
+  SUB_TITLE_PREFIX: "With Our ",
+  SUB_TITLE_SUFFIX: "Notes",
+  DESCRIPTION_ONE:
+    "Notes Library is a project dedicated to my friends to help them with their notes.",
+  DESCRIPTION_TWO: " It is your one-stop-shop for getting high quality notes.",
+};
+
+export const EXPLORE_BUTTON_TEXT = "Explore Notes";
+
+// Notes constants
+
+export const NOTES_PAGE_HEADING = "Notes";
+export const DELETE_MODAL_HEADING = "Are you absolutely sure?";
+
+export const NOTES_PAGE_DESCRIPTION_ONE = "Select the Notes That Suit You";
+export const DELETE_MODAL_DESCRIPTION =
+  "This action cannot be undone. This will permanently delete your notes and remove your notes data from our servers.";
+
+export const NOTES_PAGE_DESCRIPTION_TWO =
+  "You can start learning as soon as you download the notes.";
+
+export const NOTES_BUTTON = "View Notes";
+export const DELETE_MODAL_CANCEL_BUTTON = "Cancel";
+export const DELETE_MODAL_CONTINUE_BUTTON = "Continue";
+
+// sidebar constants
+
+export const SIDEBAR_TABS = {
+  PROFILE: "Profile",
+  NOTES: "Notes",
+  HOME: "Home",
+  SETTINGS: "Settings",
+};
+
+export const PROJECT_NAME = "Notes Library";
+
+export const LOGOUT_BUTTON_LABEL = "Logout";
