--- Frontend/src/constants/home.js
@@ -1,14 +1,0 @@
-// Home constants
-
-export const TYPED_STRINGS = ["Efficient", "Easy", "and", "Affordable"];
-
-export const TEXT_CONTENT = {
-  MAIN_TITLE: "Stay Ahead Of The Curve",
-  SUB_TITLE_PREFIX: "With Our ",
-  SUB_TITLE_SUFFIX: "Notes",
-  DESCRIPTION_ONE:
-    "Notes Library is a project dedicated to my friends to help them with their notes.",
-  DESCRIPTION_TWO: " It is your one-stop-shop for getting high quality notes.",
-};
-
-export const EXPLORE_BUTTON_TEXT = "Explore Notes";
