--- Frontend/src/constants/notes.js
@@ -1,13 +1,0 @@
-export const NOTES_PAGE_HEADING = "Notes";
-export const DELETE_MODAL_HEADING = "Are you absolutely sure?";
-
-export const NOTES_PAGE_DESCRIPTION_ONE = "Select the Notes That Suit You";
-export const DELETE_MODAL_DESCRIPTION =
-  "This action cannot be undone. This will permanently delete your notes and remove your notes data from our servers.";
-
-export const NOTES_PAGE_DESCRIPTION_TWO =
-  "You can start learning as soon as you download the notes.";
-
-export const NOTES_BUTTON = "View Notes";
-export const DELETE_MODAL_CANCEL_BUTTON = "Cancel";
-export const DELETE_MODAL_CONTINUE_BUTTON = "Continue";
