--- Frontend/src/constants/sidebar.js
@@ -1,12 +1,0 @@
-// sidebar constants
-
-export const SIDEBAR_TABS = {
-  PROFILE: "Profile",
-  NOTES: "Notes",
-  HOME: "Home",
-  SETTINGS: "Settings",
-};
-
-export const PROJECT_NAME = "Notes Library";
-
-export const LOGOUT_BUTTON_LABEL = "Logout";
